+++ apps/web/src/components/RestaurantDetails/RestaurantHeader.tsx
@@ -33,7 +33,7 @@ export default function RestaurantHeader({
   return (
     <div className="relative flex flex-col items-start justify-between">
       {/* Return button and restaurant profile */}
-      <h1 className="text-grey-100 relative flex w-full items-center justify-center py-5 text-xl font-bold">
+      <h1 className="text-grey-100 relative flex w-full items-center justify-center py-7.5 text-xl font-bold">
         <BackButton containerClassName="absolute left-0" />
         {t("restaurant profile", messages)}
       </h1>
@@ -69,8 +69,13 @@ export default function RestaurantHeader({
 
       {/* Map button */}
       <Link href={restaurant.url} className="w-full">
-        <Button className="text-1xl mt-3 flex items-center justify-center gap-1 py-5 font-semibold">
-          <Image src="/Icons/Map.svg" alt="Map button" width={20} height={20} />
+        <Button className="text-1xl mt-3 flex items-center justify-center gap-1 py-5 font-semibold hover:cursor-pointer">
+          <Image
+            src="/Icons/LocationFill.svg"
+            alt="Map button"
+            width={20}
+            height={20}
+          />
           <span>{t("map_label", messages)}</span>
         </Button>
       </Link>